requirements.txt
praw==7.8.1
google-generativeai==0.3.2
python-dotenv==1.0.1
requests==2.31.0
streamlit==1.35.0
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
pydantic==2.7.1
strip-markdown